@@ -1,12 +1,15 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { SearchResultsList } from "./index";
 
 describe("SearchResultList component test suite", () => {
   it("renders without crashing", () => {
     const div = document.createElement("div");
-    ReactDOM.render(<SearchResultsList searchResults={[]} />, div);
-    ReactDOM.unmountComponentAtNode(div);
+
+    const root = createRoot(div);
+    root.render(<SearchResultsList searchResults={[]} />);
+
+    root.unmount();
   });
 
   test.todo(
