@@ -1,19 +1,21 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { SortBar } from "./index";
 
 describe("SortBar component test suite", () => {
   it("renders without crashing", () => {
     const div = document.createElement("div");
-    ReactDOM.render(
+
+    const root = createRoot(div);
+    root.render(
       <SortBar
         sortBySongLength={() => {}}
         sortByGenre={() => {}}
         sortByPrice={() => {}}
-      />,
-      div
+      />
     );
-    ReactDOM.unmountComponentAtNode(div);
+
+    root.unmount();
   });
 
   test.todo("displays a button to sort by song length");
