@@ -1,13 +1,15 @@
 import React from "react";
-import { BrowserRouter, Route } from "react-router-dom";
-import Search from "../components/Search";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Preview from "../components/Preview";
+import Search from "../components/Search";
 import "../styles.css";
 
 const Router: React.FC = () => (
   <BrowserRouter>
-    <Route path="/" exact component={Search} />
-    <Route path="/preview/:searchResultId" component={Preview} />
+    <Routes>
+      <Route path="/" element={<Search />} />
+      <Route path="/preview/:searchResultId" element={<Preview />} />
+    </Routes>
   </BrowserRouter>
 );
 
